+++ src/ralph/utils.py
@@ -97,3 +97,23 @@ def get_dict_value_from_path(dict_: dict, path: list[str]):
         return reduce(operator.getitem, path, dict_)
     except (KeyError, TypeError):
         return None
+
+
+def set_dict_value_from_path(dict_: dict, path: list[str], value: any):
+    """Sets a nested dictionary value using an array of keys representing the path to the value."""
+
+    for key in path[:-1]:
+        dict_ = dict_.setdefault(key, {})
+    dict_[path[-1]] = value
+
+
+def remove_dict_key_from_path(dict_: dict, path: list[str]):
+    """Removes a nested dictionary key using an array of keys representing the path to the key."""
+
+    for key in path[:-1]:
+        if key not in dict_:
+            return
+        dict_ = dict_[key]
+    if path[-1] not in dict_:
+        return
+    del dict_[path[-1]]